pyproject.toml
[tool.poetry]
name = "lib"
version = "0.1.0"
description = ""
authors = ["aleksey <dazarh21.gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
redis = "^4.5.5"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
drf-yasg = "^1.21.5"
setuptools = "^67.8.0"
pandas = "^2.0.2"
flower = "^1.2.0"
celery = "5.2.7"
celery-singleton = "^0.3.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pylint = "^2.17.4"
flake8-print = "^5.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
